Task 3: standard deviation
Added a standard deviation button and a deviation button that will show
the standard deviation of the selected coin

The deviation page has 3 buttons, for the coins. Any selected coin will
display the standard deviation for the last 100 values of the coin in
the database. If there are less than 100 values then it will show for
the number of entries available